+++ example/lib/custom_grid_grouped_example.dart
@@ -23,65 +23,78 @@ class _User {
   int get hashCode => name.hashCode ^ email.hashCode;
 }
 
-class CustomGroupedExample extends StatelessWidget {
+class CustomGridGroupedExample extends StatelessWidget {
   final faker = Faker();
 
-  CustomGroupedExample({
-    Key key,
-  }) : super(key: key);
+  CustomGridGroupedExample({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     final users = List<_User>.generate(
       10,
       (i) => _User(
-        name: faker.person.name(),
+        name: faker.person.firstName(),
         email: faker.internet.email(),
       ),
-    );
-    final controller = CustomGroupController(
-      isMultipleSelection: false,
     );
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: [
         Flexible(
           fit: FlexFit.loose,
-          child: CustomGroupedCheckbox<_User>(
-            controller: controller,
+          child: CustomGroupedCheckbox<_User>.grid(
+            controller: CustomGroupController(isMultipleSelection: true),
             groupTitle: Container(
               padding: EdgeInsets.all(5.0),
               child: Text("Custom GroupedCheckbox"),
             ),
-            isScroll: true,
             itemBuilder: (ctx, index, selected, isDisabled) {
               return Card(
                 margin: EdgeInsets.only(
                   top: 5.0,
                   bottom: 5.0,
                 ),
-                child: ListTile(
-                  tileColor: selected ? Colors.green[300] : Colors.white,
-                  title: Text(users[index].name),
-                  subtitle: Text(users[index].email),
-                  trailing: Opacity(
-                    opacity: selected ? 1 : 0,
-                    child: Icon(
-                      Icons.check,
-                      color: Colors.white,
-                      size: 24,
+                child: Stack(
+                  children: [
+                    Center(
+                      child: ListTile(
+                        tileColor: Colors.white,
+                        title: Text(
+                          users[index].name,
+                          maxLines: 1,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                        subtitle: Text(
+                          users[index].email,
+                          maxLines: 1,
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                      ),
                     ),
-                  ),
+                    Positioned(
+                      left: 10,
+                      right: 10,
+                      bottom: 0,
+                      child: AnimatedContainer(
+                        duration: Duration(milliseconds: 300),
+                        color: Colors.green[300],
+                        height: selected ? 5 : 0,
+                      ),
+                    ),
+                  ],
                 ),
               );
             },
+            gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+              crossAxisCount: 3,
+            ),
             values: users,
           ),
         ),
         Builder(builder: (ctx) {
           return ElevatedButton(
             onPressed: () {
-              print(controller.selectedItem);
+              print(CustomGroupedCheckbox.of(ctx).selectedItem);
             },
             child: Text("selection"),
           );