+++ lib/src/widgets/custom_grouped_checkbox.dart
@@ -1,8 +1,8 @@
-import '../controller/custom_group_controller.dart';
-import '../common/item.dart';
 import 'package:flutter/material.dart';
 
 import '../common/custom_state_group.dart';
+import '../common/item.dart';
+import '../controller/custom_group_controller.dart';
 
 /// Signature for a function that creates a widget for a given index,isChecked and disabled, e.g., in a
 /// list.
@@ -18,37 +18,55 @@ typedef CustomIndexedWidgetBuilder = Widget Function(
 /// [groupTitle] : Text Widget that describe Title of group checkbox
 /// [itemBuilder] :  builder function  takes an index and checked state of widget
 /// [values] : list of values
-/// [itemCount] : list of initial values that you want to be selected
 /// [itemExtent] : same as [itemExtent] of [ListView]
 class CustomGroupedCheckbox<T> extends StatefulWidget {
   final CustomGroupController controller;
   final Widget? groupTitle;
   final CustomIndexedWidgetBuilder itemBuilder;
-  final int itemCount;
-  final double itemExtent;
+  final double? itemExtent;
   final List<T> values;
+  final bool _isGrid;
+  final bool isScroll;
+  final SliverGridDelegate? gridDelegate;
 
   CustomGroupedCheckbox({
     Key? key,
     required this.controller,
     this.groupTitle,
     required this.itemBuilder,
-    required this.itemCount,
     required this.values,
-    this.itemExtent = 50.0,
-  })  : assert(itemCount > 0),
+    this.isScroll = false,
+    this.itemExtent,
+  })  : this._isGrid = false,
+        this.gridDelegate = null,
+        super(key: key);
+
+  CustomGroupedCheckbox.grid({
+    Key? key,
+    required this.controller,
+    this.groupTitle,
+    SliverGridDelegate gridDelegate =
+        const SliverGridDelegateWithFixedCrossAxisCount(
+      crossAxisCount: 3,
+    ),
+    this.isScroll = false,
+    required this.itemBuilder,
+    required this.values,
+  })   : this._isGrid = true,
+        this.gridDelegate = gridDelegate,
+        this.itemExtent = 0.0,
         super(key: key);
 
   @override
   CustomGroupedCheckboxState createState() => CustomGroupedCheckboxState();
 
-  static CustomGroupedCheckboxState? of<T>(BuildContext context,
-      {bool nullOk = false}) {
-    assert(context != null);
-    assert(nullOk != null);
+  static CustomGroupController of<T>(
+    BuildContext context, {
+    bool nullOk = false,
+  }) {
     final CustomGroupedCheckboxState<T>? result =
         context.findAncestorStateOfType<CustomGroupedCheckboxState<T>>();
-    if (nullOk || result != null) return result;
+    if (nullOk || result != null) return result!.widget.controller;
     throw FlutterError.fromParts(<DiagnosticsNode>[
       ErrorSummary(
           'CustomGroupedCheckbox.of() called with a context that does not contain an CustomGroupedCheckbox.'),
@@ -76,45 +94,64 @@ class CustomGroupedCheckboxState<T>
   }
 
   @override
+  void dispose() {
+    super.dispose();
+  }
+
+  @override
   void didChangeDependencies() {
     super.didChangeDependencies();
   }
 
   @override
   Widget build(BuildContext context) {
+    final builder = (ctx, index) {
+      return ValueListenableBuilder<CustomItem<T?>>(
+        valueListenable: items[index],
+        builder: (ctx, value, child) {
+          return _ItemWidget(
+            child: widget.itemBuilder(
+              context,
+              index,
+              items[index].value.checked,
+              items[index].value.isDisabled,
+            ),
+            value: items[index].value.checked,
+            callback: (v) {
+              if (!items[index].value.isDisabled) changeSelection(index, v);
+            },
+          );
+        },
+      );
+    };
+    Widget child = ListView.builder(
+      physics: widget.isScroll
+          ? AlwaysScrollableScrollPhysics()
+          : NeverScrollableScrollPhysics(),
+      shrinkWrap: true,
+      scrollDirection: Axis.vertical,
+      itemBuilder: builder,
+      itemCount: items.length,
+      itemExtent: widget.itemExtent,
+    );
+    if (widget._isGrid) {
+      child = GridView.builder(
+        physics: widget.isScroll
+            ? AlwaysScrollableScrollPhysics()
+            : NeverScrollableScrollPhysics(),
+        gridDelegate: widget.gridDelegate!,
+        itemBuilder: builder,
+        itemCount: items.length,
+        shrinkWrap: true,
+      );
+    }
     return Column(
       children: <Widget>[
         widget.groupTitle ?? Container(),
         Expanded(
           child: ScrollConfiguration(
             behavior: ScrollBehavior(),
-            child: ListView.builder(
-              physics: NeverScrollableScrollPhysics(),
-              shrinkWrap: true,
-              scrollDirection: Axis.vertical,
-              itemBuilder: (ctx, index) {
-                return ValueListenableBuilder<CustomItem<T?>>(
-                  valueListenable: items[index],
-                  builder: (ctx, value, child) {
-                    return _ItemWidget(
-                      child: widget.itemBuilder(
-                        context,
-                        index,
-                        items[index].value.checked,
-                        items[index].value.isDisabled,
-                      ),
-                      value: items[index].value.checked,
-                      callback: (v) {
-                        if (!items[index].value.isDisabled)
-                          changeSelection(index, v);
-                      },
-                    );
-                  },
-                );
-              },
-              itemCount: items.length,
-              itemExtent: widget.itemExtent,
-            ),
+            child: child,
           ),
         ),
       ],
@@ -136,6 +173,8 @@ class CustomGroupedCheckboxState<T>
         }
       }
       items[index].value = items[index].value.copy(checked: value);
+      if (streamListValues.hasListener)
+        streamListValues.add(itemsSelections.value);
     } else {
       if (value) {
         if (itemSelected.value != null) {
@@ -148,6 +187,7 @@ class CustomGroupedCheckboxState<T>
         itemSelected.value = widget.values[index];
         items[index].value = items[index].value.copy(checked: value);
       }
+      if (streamOneValue.hasListener) streamOneValue.add(itemSelected.value);
     }
   }
 
@@ -177,7 +217,10 @@ class _ItemWidget extends StatelessWidget {
       onTap: () {
         callback!(!value!);
       },
-      child: child,
+      child: AbsorbPointer(
+        absorbing: true,
+        child: child,
+      ),
     );
   }
 }